refactor: SecretCommand and BaseCommand
